chore: sync collections and fields

+++ api/src/controllers/collections.ts
@@ -1,30 +1,49 @@
 import { Router } from 'express';
-import { ForbiddenException } from '../exceptions';
+import {ForbiddenException, UnprocessableEntityException} from '../exceptions';
 import { respond } from '../middleware/respond';
 import { validateBatch } from '../middleware/validate-batch';
 import { CollectionsService, MetaService } from '../services';
 import { Item } from '../types';
 import asyncHandler from '../utils/async-handler';
+import {isTemplate} from '../database/helpers/is-template';
+import {isMaster} from '../database/helpers/is-master';
+import {getChildDatabases} from '../database/helpers/get-child-databases';
+import { Knex } from 'knex';
 
 const router = Router();
 
 router.post(
 	'/',
 	asyncHandler(async (req, res, next) => {
-		const collectionsService = new CollectionsService({
-			accountability: req.accountability,
-			schema: req.schema,
-			knex: req.knex,
-		});
+		if (isTemplate(req.knex) || isMaster(req.knex)) {
+			let databases: Knex[] = [];
 
-		if (Array.isArray(req.body)) {
-			const collectionKey = await collectionsService.createMany(req.body);
-			const records = await collectionsService.readMany(collectionKey);
-			res.locals.payload = { data: records || null };
+			if (isMaster(req.knex)) {
+				databases.push(req.knex);
+			} else {
+				databases = databases.concat(await getChildDatabases());
+			}
+
+
+			for (const database of databases) {
+				const collectionsService = new CollectionsService({
+					accountability: req.accountability,
+					schema: req.schema,
+					knex: database,
+				});
+
+				if (Array.isArray(req.body)) {
+					const collectionKey = await collectionsService.createMany(req.body);
+					const records = await collectionsService.readMany(collectionKey);
+					res.locals.payload = { data: records || null };
+				} else {
+					const collectionKey = await collectionsService.createOne(req.body);
+					const record = await collectionsService.readOne(collectionKey);
+					res.locals.payload = { data: record || null };
+				}
+			}
 		} else {
-			const collectionKey = await collectionsService.createOne(req.body);
-			const record = await collectionsService.readOne(collectionKey);
-			res.locals.payload = { data: record || null };
+			throw new UnprocessableEntityException(`Database structure in child database is readonly`);
 		}
 
 		return next();
@@ -82,23 +101,37 @@ router.get(
 router.patch(
 	'/:collection',
 	asyncHandler(async (req, res, next) => {
-		const collectionsService = new CollectionsService({
-			accountability: req.accountability,
-			schema: req.schema,
-			knex: req.knex,
-		});
+		if (isTemplate(req.knex) || isMaster(req.knex)) {
+			let databases: Knex[] = [];
 
-		await collectionsService.updateOne(req.params.collection, req.body);
-
-		try {
-			const collection = await collectionsService.readOne(req.params.collection);
-			res.locals.payload = { data: collection || null };
-		} catch (error: any) {
-			if (error instanceof ForbiddenException) {
-				return next();
+			if (isMaster(req.knex)) {
+				databases.push(req.knex);
+			} else {
+				databases = databases.concat(await getChildDatabases());
 			}
 
-			throw error;
+			for (const database of databases) {
+				const collectionsService = new CollectionsService({
+					accountability: req.accountability,
+					schema: req.schema,
+					knex: database,
+				});
+
+				await collectionsService.updateOne(req.params.collection, req.body);
+
+				try {
+					const collection = await collectionsService.readOne(req.params.collection);
+					res.locals.payload = { data: collection || null };
+				} catch (error: any) {
+					if (error instanceof ForbiddenException) {
+						return next();
+					}
+
+					throw error;
+				}
+			}
+		} else {
+			throw new UnprocessableEntityException(`Database structure in child database is readonly`);
 		}
 
 		return next();
@@ -109,13 +142,27 @@ router.patch(
 router.delete(
 	'/:collection',
 	asyncHandler(async (req, res, next) => {
-		const collectionsService = new CollectionsService({
-			accountability: req.accountability,
-			schema: req.schema,
-			knex: req.knex,
-		});
+		if (isTemplate(req.knex) || isMaster(req.knex)) {
+			let databases: Knex[] = [];
 
-		await collectionsService.deleteOne(req.params.collection);
+			if (isMaster(req.knex)) {
+				databases.push(req.knex);
+			} else {
+				databases = databases.concat(await getChildDatabases());
+			}
+
+			for (const database of databases) {
+				const collectionsService = new CollectionsService({
+					accountability: req.accountability,
+					schema: req.schema,
+					knex: database,
+				});
+
+				await collectionsService.deleteOne(req.params.collection);
+			}
+		} else {
+			throw new UnprocessableEntityException(`Database structure in child database is readonly`);
+		}
 
 		return next();
 	}),
